+++ R/sparegcv.R
@@ -27,6 +27,8 @@
 #'         \code{"class"} (misclassification error) and
 #'         \code{"1-auc"} (one minus area under the ROC curve) both just for
 #'         binomial family.
+#' @param avg_type type of averaging the marginal models; either on link (default)
+#'        or on response level. This is used in computing the validation measure.
 #' @param parallel assuming a parallel backend is loaded and available, a
 #'        logical indicating whether the function should use it in parallelizing the
 #'        estimation of the marginal models. Defaults to FALSE.
@@ -36,21 +38,27 @@
 #' \itemize{
 #'  \item \code{betas} p x  \code{max(nummods)} sparse matrix of class
 #'   \code{'\link[Matrix:dgCMatrix-class]{Matrix::dgCMatrix}'} containing the
-#'   standardized coefficients from each marginal model
+#'   standardized coefficients from each marginal model computed with the spar
+#'   algorithm on the whole training data.
 #'  \item \code{intercepts} used in each marginal model, vector of length \code{max(nummods)}
+#'    computed with the spar algorithm on the whole training data.
 #'  \item \code{scr_coef} p-vector of coefficients used for screening for standardized predictors
 #'  \item \code{inds} list of index-vectors corresponding to variables kept after
 #'  screening in each marginal model of length  \code{max(nummods)}
 #'  \item \code{RPMs} list of projection matrices used in each marginal model of length \code{max(nummods)}
-#'  \item \code{val_sum} a \code{data.frame} with CV results (mean and sd validation measure and mean number of active variables) for each element of nus and nummods
+#'  \item \code{val_res} a \code{data.frame} with CV results for each fold and for each element of nus and nummods
 #'  \item \code{nus} vector of \eqn{\nu}'s considered for thresholding
 #'  \item \code{nummods} vector of numbers of marginal models considered for validation
+#'  \item \code{family}  a \link[stats]{family}  object used for the marginal generalized linear model
+#'  \item \code{measure} character, type of validation measure used
+#'  \item \code{avg_type} character, averaging type for computing the validation measure
+#'  \item \code{rp} an object of class \code{'randomprojection'}
+#'  \item \code{screencoef} an object of class \code{'screeningcoef'}
+#'  \item \code{model} an object of class \code{'sparmodel'}
 #'  \item \code{ycenter} empirical mean of initial response vector
 #'  \item \code{yscale} empirical standard deviation of initial response vector
 #'. \item \code{xcenter} p-vector of empirical means of initial predictor variables
 #'  \item \code{xscale} p-vector of empirical standard deviations of initial predictor variables
-#'  \item \code{rp} an object of class \code{'randomprojection'}
-#'  \item \code{screencoef} an object of class \code{'screeningcoef'}
 #' }
 #' @examples
 #' \donttest{
@@ -74,8 +82,9 @@
 #' @export
 spar.cv <- function(x, y, family = gaussian("identity"), model = spar_glmnet(),
                     rp = NULL, screencoef = NULL, nfolds = 10,
-                    nnu = 20, nus = NULL,
-                    nummods = c(20), measure = c("deviance","mse","mae","class","1-auc"),
+                    nnu = 20, nus = NULL, nummods = c(20),
+                    measure = c("deviance","mse","mae","class","1-auc"),
+                    avg_type = c("link","response"),
                     parallel = FALSE, seed = NULL, ...) {
   # Set up and checks ----
   n <- length(y)
@@ -89,7 +98,7 @@ spar.cv <- function(x, y, family = gaussian("identity"), model = spar_glmnet(),
   }
 
   measure <- match.arg(measure)
-
+  avg_type <- match.arg(avg_type)
   # Ensure back compatibility ----
   args <- list(...)
   arg_list <- check_and_set_args(args, x, y, family, model,
@@ -105,6 +114,7 @@ spar.cv <- function(x, y, family = gaussian("identity"), model = spar_glmnet(),
                             nnu = nnu, nus = nus,
                             nummods = nummods,
                             measure = measure,
+                            avg_type = avg_type,
                             inds = NULL, RPMs = NULL,
                             parallel = parallel,
                             seed = seed)
@@ -125,6 +135,7 @@ spar.cv <- function(x, y, family = gaussian("identity"), model = spar_glmnet(),
       RPMs = SPARres$RPMs,
       nummods = nummods,
       measure = measure,
+      avg_type = avg_type,
       parallel = parallel,
       seed = seed)
     val_res <- rbind(val_res,
@@ -138,7 +149,7 @@ spar.cv <- function(x, y, family = gaussian("identity"), model = spar_glmnet(),
               val_res = val_res,
               nus = SPARres$nus, nummods=nummods,
               family = family,
-              measure = measure,
+              measure = measure, avg_type = avg_type,
               rp = rp, screencoef = screencoef,
               model = model,
               ycenter = SPARres$ycenter, yscale = SPARres$yscale,